Disable -Xjnigreflimit.
This was supposed to help find bugs, but in practice (unlike the
locale reference limit and the pinned array limit) doesn't. What it
does do is cause well-behaved code to fail if its allocation rate
is high. This has caused trouble for both audio and graphics code
running on eng builds.

Bug: 7903975
Change-Id: If475cb51c9cab13270a83a60d6d0aecfab758e88